refactor label printing in its own file
and cleanup imports
